[software] ExportAnimatedCamera: Drop all but cosmetic changes
This commit drops the changes that were made to `main_exportAnimatedCamera`
in the previous commits, except for the cosmetic changes and, in
particular, the 4-space indentation.

These changes need to be dropped as they are not meant to be added with
the distortion calibration, and may cause errors when exporting the
animated cameras for which the distortion was calibrated.

It additionally harmonizes the comments in the code.